JPA met spring Data toegepast op brouwers oefening. + paginering brouwers

+++ brouwers/src/main/java/be/vdab/services/BrouwerServiceImpl.java
@@ -3,6 +3,9 @@ package be.vdab.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.annotation.Isolation;
@@ -23,34 +26,37 @@ class BrouwerServiceImpl implements BrouwerService {
 	@Override
 	@Transactional(readOnly = false, isolation = Isolation.READ_COMMITTED) 
 	public void create(Brouwer Brouwer) {
-		brouwerDAO.create(Brouwer);
+		brouwerDAO.save(Brouwer);
 	}
   	
 	@Override 
 	public void read(long id) {
-		brouwerDAO.read(id);
+		brouwerDAO.findOne(id);
 	}
 	
 	@Override
 	@Transactional(readOnly = false, isolation = Isolation.READ_COMMITTED) 
 	public void update(Brouwer Brouwer) {
-		brouwerDAO.update(Brouwer);
+		brouwerDAO.save(Brouwer);
 	}
 	
 	@Override
 	@Transactional(readOnly = false, isolation = Isolation.READ_COMMITTED) 
 	public void delete(long id) {
-		brouwerDAO.delete(id);
+		Brouwer brouwer  = brouwerDAO.findOne(id);
+		if (brouwer != null) {
+		  brouwerDAO.delete(id);
+		}
 	}
 	
 	@Override
-	public List<Brouwer> findAll() {
-		return brouwerDAO.findAll();
+	public Page<Brouwer> findAll(Pageable pageable) {
+		return brouwerDAO.findAll(pageable);
 	} 
   
 	@Override
 	public List<Brouwer> findByNaam(String beginNaam) {
-		return brouwerDAO.findByNaam(beginNaam);
+		return brouwerDAO.findByNaamStartingWith(beginNaam);
 	}
 
 }
